Refactor data init to merge redundant ObjectStore classes
