--- app/entrypoints/initialize_all_data.sh
@@ -1,15 +1,0 @@
-#!/bin/env bash
-
-set -o pipefail
-set -e
-
-# Create data folders on persistent volume and symlink to expected paths
-SCRIPT_DIR="$(dirname "$(readlink -f "$0")")"
-cd "${SCRIPT_DIR}"
-
-bash initialize_data_dirs.sh
-
-# Verify and download missing and invalid files
-python initialize_data.py download
-
-echo "Data initialization complete."
