+++ app/host/object_store.py
@@ -6,23 +6,28 @@ import io
 import os
 import json
 from uuid import uuid4
+import hashlib
+import re
 from host.log import get_logger
 logger = get_logger(__name__)
 
 
 class ObjectStore:
-    def __init__(self) -> None:
+    def __init__(self, conf: dict = {}) -> None:
         '''Initialize S3 client'''
         # Randomize base path if not provided to avoid accidental overwrite of existing objects
         random_base_path = str(uuid4())
         self.config = {
             'endpoint-url': os.getenv("S3_ENDPOINT_URL", ""),
             'region-name': os.getenv("S3_REGION_NAME", ""),
-            'aws_access_key_id': os.getenv("AWS_S3_ACCESS_KEY_ID"),
-            'aws_secret_access_key': os.getenv("AWS_S3_SECRET_ACCESS_KEY"),
-            'bucket': os.getenv("S3_BUCKET"),
+            'aws_access_key_id': os.getenv("AWS_S3_ACCESS_KEY_ID", ""),
+            'aws_secret_access_key': os.getenv("AWS_S3_SECRET_ACCESS_KEY", ""),
+            'bucket': os.getenv("S3_BUCKET", ""),
             'base_path': os.getenv("S3_BASE_PATH", f'''/{random_base_path}'''),
         }
+        # Override default config values with input dict
+        for key, val in conf.items():
+            self.config[key] = val
         self.bucket = self.config['bucket']
         self.base_path = self.config['base_path']
         self.client = None
@@ -100,13 +105,21 @@ class ObjectStore:
             object_name=key)
         return response.stream(32 * 1024)
 
-    def download_object(self, path="", file_path=""):
+    def download_object(self, path="", file_path="", version_id=""):
         path = path.strip('/')
-        self.client.fget_object(
-            bucket_name=self.bucket,
-            object_name=path,
-            file_path=file_path,
-        )
+        if version_id:
+            self.client.fget_object(
+                bucket_name=self.bucket,
+                object_name=path,
+                file_path=file_path,
+                version_id=version_id,
+            )
+        else:
+            self.client.fget_object(
+                bucket_name=self.bucket,
+                object_name=path,
+                file_path=file_path,
+            )
 
     def delete_directory(self, root_path):
         delete_object_list = map(
@@ -121,6 +134,15 @@ class ObjectStore:
             delete_object_list=delete_object_list)
         for error in errors:
             logger.error("Error deleting object: ", error)
+
+    def get_directory_objects(self, root_path):
+        objects = self.client.list_objects(
+            bucket_name=self.bucket,
+            prefix=root_path,
+            include_version=True,
+            recursive=True,
+        )
+        return [obj for obj in objects]
 
     def list_directory(self, root_path, recursive=True):
         objects = self.client.list_objects(
@@ -165,3 +187,50 @@ class ObjectStore:
                 object_name=dst_path,
                 source=CopySource(self.bucket, object_name))
             logger.debug(f'''Copied object "{result.object_name}" ({result.version_id})''')
+
+    def md5_checksum(self, file_path):
+        '''https://stackoverflow.com/a/58239738'''
+        m = hashlib.md5()
+        with open(file_path, 'rb') as fh:
+            for data in iter(lambda: fh.read(1024 * 1024), b''):
+                m.update(data)
+        hexdigest = m.hexdigest()
+        logger.debug(f'calculated md5 checksum: {hexdigest}')
+        return hexdigest
+
+    def etag_checksum(self, file_path, etag_parts=1, file_size=0):
+        '''https://stackoverflow.com/a/58239738'''
+        md5s = []
+        min_chunk_size = 16 * 1024**2
+        chunk_size = int(file_size / etag_parts)
+        if etag_parts == 1:
+            chunk_size = file_size
+        elif chunk_size < min_chunk_size:
+            chunk_size = min_chunk_size
+        chunk_size_mib = int(chunk_size / 1024**2)
+        file_size_mib = int(file_size / 1024**2)
+        logger.debug(f"chunk_size is {chunk_size_mib} MiB for file size {file_size_mib} bytes"
+                     "(etag parts: {etag_parts})")
+        with open(file_path, 'rb') as fh:
+            for data in iter(lambda: fh.read(chunk_size), b''):
+                md5s.append(hashlib.md5(data).digest())
+        digests_md5 = hashlib.md5(b''.join(md5s))
+        etag_checksum = f'{digests_md5.hexdigest()}-{etag_parts}'
+        logger.debug(f'calculated etag: {etag_checksum}')
+        return etag_checksum
+
+    def etag_compare(self, file_path, etag_source, file_size):
+        '''https://stackoverflow.com/a/58239738'''
+        etag_source = etag_source.strip('"')
+        etag_local = ''
+        if '-' in etag_source:
+            etag_parts = int(re.search(r'^.+-([0-9]+$)', etag_source).group(1))
+            etag_local = self.etag_checksum(file_path, etag_parts=etag_parts, file_size=file_size)
+        elif '-' not in etag_source:
+            etag_local = self.md5_checksum(file_path)
+        if etag_source == etag_local:
+            return True
+        else:
+            logger.warning(f'    source etag checksum: {etag_source}')
+            logger.warning(f'     local etag checksum: {etag_local}')
+        return False
